Remove file clone
The file object passed by gulp is now a wrapped node file object. Clone
seems to be tripping over the path setting resulting in .scss files
rather than .css files. The content of the files is, as expected, css.

This fix removes the file cloning and instead just emits the incoming
file with modified attributes.